bugfix: only show cancle journey, when it's the own journey

+++ lib/widgets/request_and_offer_card.dart
@@ -16,8 +16,10 @@ class RequestAndOfferCard extends StatelessWidget {
   bool _isCurrentUserAccepted = false;
   bool _isCurrentUserDeclined = false;
   final void Function() refreshParent;
+  final bool showDeleteCloseButtons;
 
-  RequestAndOfferCard(this.journey, this.refreshParent, [this.currentUserId]) {
+  RequestAndOfferCard(this.journey, this.refreshParent,
+      [this.currentUserId, this.showDeleteCloseButtons = true]) {
     if (journey.pendingUserIds != null &&
         journey.pendingUserIds.contains(currentUserId)) {
       _isCurrentUserPending = true;
@@ -83,8 +85,12 @@ class RequestAndOfferCard extends StatelessWidget {
                     _isCurrentUserPending ||
                         _isCurrentUserAccepted ||
                         _isCurrentUserDeclined);
-              } else {
+              } else if (showDeleteCloseButtons) {
                 return DeleteCloseJourneyButtonRow(journey, refreshParent);
+              } else {
+                // yes, this is an empty widget. It is needed as otherwise there
+                // would be an error, because a widget has to be returned to the column
+                return SizedBox.shrink();
               }
             }())
           ],
